Fix: dashboard active states.
There were so many links that had to be changed. Initially there was no need to change any real codes. It's just the links and some other stuffs like, .html instead of .php, index.html instead of dashboard/dashbard.html, were the reasons which caused it to not work as expected.

resume.html had only the main container, there were no html,head,body tags in it, which were necessary to run a web document.